api/requirements.txt
mechanize
requests
beautifulsoup4
python-dotenv
flask